requirements/base.txt
choice-enum==0.9
Django<2.0
django-anymail==0.10
django-bootstrap3==8.2.3
django-extensions==1.8.0
html5lib<1.0
mysqlclient==1.3.10
Pillow==4.1.1
python-dateutil==2.6.0
pytz==2017.2
requests==2.18.1
six==1.10.0
urllib3==1.21.1
Unidecode==0.4.21

-e git+https://github.com/dakrauth/dbgutils.git#egg=dbgutils
-e git+https://github.com/dakrauth/pagination.git#egg=pagination
